requirements.txt
pyrogram[fast]
pillow
